move InterestDatabase to interfaces package
