+++ app/src/main/java/com/github/polybooks/core/database/interfaces/Database.kt
@@ -1,7 +1,5 @@
 package com.github.polybooks.core.database.interfaces
 
-import com.github.polybooks.core.database.InterestDatabase
-
 /**
  * A Database aggregates the functionality of Book/Sale/Interest Databases in one place
  * */
